kbuild: Generate versions for exported symbols
Up to now, we had a way to store the checksums associated with the
exported symbols, but they were not filled in yet. This is done
with this patch, using the linker to actually do that for us.
  
The comment added with this patch explains what magic exactly is going
on.