[PATCH] alpha numa setup_memory leaves meaningless {min,max}_low_pfn
Currently, after the numa version of setup_memory, {min,max}_low_pfn are
set to the values used in the calculation of the last potential node
rather than anything meaningful to the system. On side-effect is that if
there is no memory in the last node considered, min_low_pfn is left at
~0UL resulting in initrd being disabled as "overwritten".

This patch makes setup_memory_node use local node_{min,max}_pfn
variables in computing memory for the node rather than the global
{min,max}_low_pfn. For nodes with memory, the limits are propagated to
the globals.

/jeff