[ARM PATCH] 1380/1: Big-Endian support for jiffies
Patch from Deepak Saxena

jiffies is defined as pointing to jiffies_64 for backwards comptability (I think, not 100% sure why we don't just use jiffies_64 everywhere?). On BE systems, we need to point to the upper 4 bytes of jiffies_64 or we will never get past calibrate_delay. This is what is done in PPC, SH, and other big endian architectures.