[PATCH] ips: missing reboot notifier and Mode Sense P8
This adds back the reboot_notifier hook that was accidentally removed in
the last set of patches.

It also adds support handle a Mode Page 8 command so the scsi layer
won't complain anymore.  And the movement to using ->device to get scsi
numbers resulted in 2 typos in debug code.  These are also fixed.