[atm nicstar] fix incorrect traffic class assumption
I have encountered a bug in the nicstar ATM driver for linux.
If you open a CBR TX only connection on a specific vpi/vci and later open a RX 
only connection on the same vpi/vci, the RX connection will overwrite the 
pointer to the SCQ of the TX connection. This changes the cell rate of the TX 
channel and what is worse is that when the TX connection is closed we get a 
segmentationfault and the TX part of the vpi/vci remains reserved.

The bug in the driver is that if the opened channel is not TX CBR, the driver 
assumes it is TX UBR. I have attached a patch that adds a check for TX UBR. 
The patch is against RedHat kernel 2.4.18-3. I have checked linux vanilla 
kernels 2.4.19 and 2.5.49 and not found a fix.