kbuild: Add dependency info to modules
During postprocessing, we get the information which modules we depend on 
(i.e. which modules export symbols we need to import) for free, so we
record it in .modinfo. This may allow us to remove the "depmod" program
at a later time, though for now it remains working as usual.