do_mounts: Simplify logic for ramdisk from floppy.
Currently, we would try to read in a ramdisk image from floppy, if
o root device is /dev/fd*
o "load_ramdisk=1" on the command line
o CONFIG_BLK_DEV_INITRD is not set, or "noinitrd" on the command line.

Relax the last restriction, which only makes things more complicated for
no reason, and changes behavior depending on an unrelated config option.