[PATCH] USB: KB Gear USB Tablet Drivers
Attached are drivers for the KB Gear JamStudio Tablet. There are two
files, one is against 2.4.20, the other against 2.5.62. I'm hoping this
isn't too late in the 2.4.21-pre stages to get included (or the 2.5, for
that matter =). This driver Works For Me, on both 2.4.20 and 2.5.62.

Anyway, as usual, comments, complaints, and patches are more than
welcome.